backend/requirements.txt
asgiref==3.9.1
coverage==7.9.2
Django==4.2.23
django-cors-headers==3.14.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.4.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
